requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.3.1
pymupdf==1.24.9
unstructured==0.15.7
tiktoken==0.7.0

faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
langchain==0.2.14
langchain-community==0.2.11

transformers==4.44.2
accelerate==0.34.2

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1

streamlit==1.37.1
